agent-mocked-tools/requirements.txt
strands-agents[openai]
strands-agents-tools
boto3
bedrock-agentcore
aws-opentelemetry-distro
